+++ data-raw/make_data.R
@@ -8,6 +8,8 @@ psid    <- utils::read.csv("data-raw/psid_wages_panel.csv", stringsAsFactors = F
 twins   <- utils::read.csv("data-raw/twins.csv", stringsAsFactors = FALSE)
 castle  <- utils::read.csv("data-raw/castle_panel.csv",    stringsAsFactors = FALSE)
 divorce <- utils::read.csv("data-raw/divorce_panel.csv",   stringsAsFactors = FALSE)
+minimum_wage <- utils::read.csv("data-raw/minimum_wage_panel.csv",
+                                stringsAsFactors = FALSE)
 fscore  <- utils::read.csv("data-raw/f_score_panel.csv",   stringsAsFactors = FALSE)
 
 if (!dir.exists("data")) dir.create("data")
@@ -16,6 +18,7 @@ save(psid,    file = "data/psid.rda",    compress = "xz")
 save(twins,   file = "data/twins.rda",   compress = "xz")
 save(castle,  file = "data/castle.rda",  compress = "xz")
 save(divorce, file = "data/divorce.rda", compress = "xz")
+save(minimum_wage, file = "data/minimum_wage.rda", compress = "xz")
 save(fscore,  file = "data/fscore.rda",  compress = "xz")
 
 message("built: ", paste(list.files("data"), collapse = ", "))